requirements.txt
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.1
nltk==3.8.1
textblob==0.17.1
praw==7.7.1
tweepy==4.14.0
yfinance==0.2.31
transformers==4.34.0
torch==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0 